script to shorten filenames of web assets
since SPIFFS only allows 32-character paths, we need to shorten the
filenames of the web assets.  When doing so, we also need to update
the references to the files in the code.

To test this, create a simple test server which will only serve the
files in the build_shortened directory.  This way, we can test the
web interface without having to upload the files to the ESP32.

Finally, set DCONFIG_ASYNC_TCP_RUNNING_CORE=0 to try and
eliminate the lurches in the stepper motor when the web server
is active.

Also, don't call /text very often, since apparently it is a little
taxing.  Instead, only call it on load and then on demand.